Add files via upload
This report provides a comprehensive description of all the details of the Blackboard application. The content of the report “Index” has been changed and a detailed description of the Blackboard application has been added in the first chapter. In the first chapter, we have added an introduction to the application, its purpose, the people involved in it, the main tasks and requirements of the Blackboard application, and finally the conclusion of the first chapter.

In the second chapter, we have added some very important points in the second screen: the presence of the Blackboard application, the challenges that the Blackboard application solves, the opinion of one group, a comparison between two applications with similar functions, the pros and cons, and finally the conclusion.

In the third chapter, we have added the interface for the application users, the data model used by the application, the application database and the supported programming languages. In addition, images have been included that visually display the interface design to provide a complete and clear description of all the basic elements of the application, and finally the conclusion at the end.
The report concludes with the final conclusions and references for the Blackboard application.