Add files via upload
In the third chapter of the Latex file, all the details of the Blackboard application are explained comprehensively. The chapter includes an explanation of the user interfaces, the data models used in the application, its database, and the programming languages ​​it relies on. In addition, the references added in the file are used